feat: Moved logger repository
